test: use new data see to run cli tests
