+++ cli/.test.env
@@ -1,4 +1,4 @@
-TRANSFORM_PROJECT=https://cube-creator.lndo.site/project/cli-test
+TRANSFORM_PROJECT=https://cube-creator.lndo.site/cube-project/ubd
 
 AUTH_RUNNER_CLIENT_ID=cube-creator-runner
 AUTH_RUNNER_ISSUER=https://keycloak.zazukoians.org/realms/zazuko-dev
