+++ cli/test/lib/projects.test.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai'
 import { Hydra } from 'alcaeus/node'
 import env from '@cube-creator/core/env'
 import { ProjectIterator } from '../../lib/project'
-import { insertTestData } from '../support/testData'
+import { insertTestData } from '@cube-creator/testing/lib'
 import { setupEnv } from '../support/env'
 import { Table } from '@rdfine/csvw'
 import { log } from '../support/logger'
@@ -17,7 +17,7 @@ describe('lib/projects', function () {
 
   before(async () => {
     setupEnv()
-    await insertTestData()
+    await insertTestData('fuseki/sample-ubd.trig')
   })
 
   beforeEach(() => {
@@ -27,7 +27,7 @@ describe('lib/projects', function () {
   describe('ProjectIterator', () => {
     it('streams csv table objects from project', async () => {
       // given
-      const iteratorStream = new ProjectIterator({ projectUri: '/project/cli-test', log, variables })
+      const iteratorStream = new ProjectIterator({ projectUri: '/cube-project/ubd', log, variables })
 
       // when
       const results: Table[] = []
@@ -36,8 +36,8 @@ describe('lib/projects', function () {
       }
 
       // then
-      expect(results).to.have.length(1)
-      expect(results[0].id.value).to.match(new RegExp('/project/cli-test/mapping/table/foo/csvw$'))
+      expect(results).to.have.length(2)
+      expect(results[0].id.value).to.match(new RegExp('/project/ubd/csv-mapping/table-\\w+/csvw$'))
       expect(results[0].dialect?.quoteChar).to.equal('"')
       expect(results[0].dialect?.delimiter).to.equal(',')
       expect(results[0].dialect?.header).to.equal(true)
@@ -45,7 +45,7 @@ describe('lib/projects', function () {
 
     it('sets cube URI as pipeline variable "graph"', async () => {
       // given
-      const iteratorStream = new ProjectIterator({ projectUri: '/project/cli-test', log, variables })
+      const iteratorStream = new ProjectIterator({ projectUri: '/cube-project/ubd', log, variables })
 
       // when
       await new Promise((resolve, reject) => {
