+++ packages/testing/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@cube-creator/testing",
+  "version": "0.0.0",
+  "private": true,
+  "dependencies": {
+    "@cube-creator/core": "*",
+    "@rdfjs/formats-common": "^2.1.0",
+    "@tpluscode/rdf-ns-builders": "^0.4.0",
+    "@tpluscode/sparql-builder": "^0.3.9",
+    "commander": "^4.1.1",
+    "debug": "^4.1.1",
+    "sparql-http-client": "^2.2.2"
+  }
+}
